feat(storyblok-ui): ship Visual Editor proxy as auto-loaded plugin
Move the Storyblok Visual Editor locale-routing logic from the example
into @graphcommerce/storyblok-ui as a function plugin with ifConfig:
'storyblok'. Now any project depending on @graphcommerce/storyblok-ui
gets Visual Editor locale routing automatically — no manual plugins/
file in the project.

The example's plugins/LocaleRedirectProxy.ts now only handles the
generic Accept-Language root redirect (project-level opt-in).
