+++ examples/magento-storyblok/plugins/LocaleRedirectProxy.ts
@@ -9,7 +9,7 @@ export const config: PluginConfig = {
 }
 
 const LOCALES = storefront.map((s) => s.locale)
-const DEFAULT_LOCALE = storefront.find((s) => s.defaultLocale)?.locale || 'en'
+const DEFAULT_LOCALE = storefront.find((s) => s.defaultLocale)?.locale || LOCALES[0]
 
 /**
  * Locale detection is disabled in next.config.ts (`localeDetection: false`) to prevent the
@@ -17,24 +17,11 @@ const DEFAULT_LOCALE = storefront.find((s) => s.defaultLocale)?.locale || 'en'
  * re-implements the same behaviour for regular visitors: on the root path, parse Accept-Language
  * and redirect to the preferred locale if it differs from the default.
  *
- * For the Storyblok Visual Editor, the `_storyblok_lang` param determines which locale to show.
- * When it differs from the current URL locale, the proxy redirects to the correct locale path.
+ * Storyblok-specific Visual-Editor locale routing lives in
+ * `@graphcommerce/storyblok-ui/plugins/StoryblokVisualEditorProxy.ts`.
  */
 export const proxy: FunctionPlugin<typeof proxyType> = (prev, request) => {
-  if (request.nextUrl.searchParams.has('_storyblok')) {
-    const sbLang = request.nextUrl.searchParams.get('_storyblok_lang')
-    const currentLocale = request.nextUrl.locale || DEFAULT_LOCALE
-    const targetLocale = sbLang === 'default' || !sbLang ? DEFAULT_LOCALE : sbLang
-
-    if (LOCALES.includes(targetLocale) && targetLocale !== currentLocale) {
-      const url = request.nextUrl.clone()
-      url.locale = targetLocale
-      return NextResponse.redirect(url)
-    }
-
-    return prev(request)
-  }
-
+  if (request.nextUrl.searchParams.has('_storyblok')) return prev(request)
   if (request.nextUrl.pathname !== '/') return prev(request)
   if (request.cookies.has('NEXT_LOCALE')) return prev(request)
